requirements.txt
flask
redis
mysql-connector-python-rf
GPSPhoto
exifread
Pillow
piexif
google-api-python-client
google-cloud-automl